Removed the k-nearest neighbours search in the KDTree class. Fixed a portability problem in the BoxCoxEvaluationImplementation class. Fixed a typo in the documentation of the NumericalMathFunctionImplementation class. Added a missing OT_API directive in several classes. Added the ConvertToRGB() and ConvertToRGBA() methods to the Drawable class. Added the ScanColorCode(), ConvertToRGB() and ConvertToRGBA() methods to the DrawableImplementation class. Reverted the Polygon class to its previous implementation. Added the Polygonarray class. Based the implementation of the drawMarginal() method of the FieldImplementation class on the PolygonArray class, leading to a huge speed improvement. Added a draw() method to the Field and FieldImplementation classes. Used memcpy in several places in the Matrix and Tensor related classes. Applied Denis Barbier's patch to disambiguate the call to many mathematical functions. Changed the behaviour of the viewer for the grid and the legend of clouds with 'none' point style.
distcheck ok.

+++ lib/src/Uncertainty/Distribution/Dirac.cxx
@@ -214,7 +214,7 @@ NumericalPoint Dirac::getStandardMoment(const UnsignedInteger n) const
 {
   const UnsignedInteger dimension(getDimension());
   NumericalPoint result(dimension);
-  for (UnsignedInteger i = 0; i < dimension; ++i) result[i] = std::pow(point_[i], n);
+  for (UnsignedInteger i = 0; i < dimension; ++i) result[i] = std::pow(point_[i], static_cast<int>(n));
   return result;
 }
