+++ lib/src/Uncertainty/Distribution/Histogram.cxx
@@ -307,7 +307,7 @@ NumericalPoint Histogram::getStandardMoment(const UnsignedInteger n) const
   for (UnsignedInteger i = 0; i < size; ++i)
     {
       const NumericalScalar x(xPrec + collection_[i].getWidth() * factor);
-      value += (std::pow(x, n + 1) - std::pow(xPrec, n + 1)) * collection_[i].getHeight();
+      value += (std::pow(x, static_cast<int>(n + 1)) - std::pow(xPrec, static_cast<int>(n + 1))) * collection_[i].getHeight();
       xPrec = x;
     }
   value /= (n + 1) * factor;
